lab I2

+++ azheganova.faina/I2/main.cpp
@@ -27,6 +27,7 @@ int main()
     std::cout << "Not a number";
     return 1;
   }
+  
   int* dynamicarray = new int[n];
   for (size_t i = 0; i < n; i++)
   {
@@ -34,7 +35,7 @@ int main()
   }
   try
   {
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
     {
       std::cout << dynamicarray[i] << " ";
     }
